Fix license issues
